+++ release_notes.md
@@ -3,9 +3,15 @@
 <!-- Please add your release notes in the following format:
 - My change description (#PR)
 -->
-
+- Update Python Worker Version to [4.9.0](https://github.com/Azure/azure-functions-python-worker/releases/tag/4.9.0)
+- Update Java Worker Version to [2.8.0](https://github.com/Azure/azure-functions-java-worker/releases/tag/2.8.0)
+- Track ApplicationInsights ingestion service response by subscribing to transmission event [PR](https://github.com/Azure/azure-functions-host/pull/8828)
 - Update ApplicationInsights packages to 2.21.0 (#8838)
 - Support model-type bindings for out-of-proc workers (#8815)
+- Fix the duplicate log issue with Java ApplicationInsights agent
+- Update Node.js Worker Version to [3.5.2](https://github.com/Azure/azure-functions-nodejs-worker/releases/tag/v3.5.2)
+- Making feature flag for worker indexing opt-out instead of opt-in (#8864)
+
 
 **Release sprint:** Sprint 132
 [ [bugs](https://github.com/Azure/azure-functions-host/issues?q=is%3Aissue+milestone%3A%22Functions+Sprint+132%22+label%3Abug+is%3Aclosed) | [features](https://github.com/Azure/azure-functions-host/issues?q=is%3Aissue+milestone%3A%22Functions+Sprint+132%22+label%3Afeature+is%3Aclosed) ]